requirements.txt
streamlit
gspread
oauth2client
google-genai
requests
beautifulsoup4
